+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 
 - **[BC BREAK]** Use `ResponseDefinition` instead of obsolete `Result`
+- **[BC BREAK]** Remove `Responder` interface
 
 ## [0.5.0] - 2021-07-04